Removed MAUI sample
Restored WPF Sample

Updated to latest SDK

Updated packages
