+++ LoDaTek.AzureDevOps.Services.Client/LoDaTek.AzureDevOps.Services.Client.csproj
@@ -22,8 +22,8 @@
 	</ItemGroup>
 	
 	<ItemGroup Condition=" $(TargetFramework.StartsWith('netstandard')) ">
-		<PackageReference Include="System.Net.Http.Json" Version="10.0.5" />
-		<PackageReference Include="System.Text.Json" Version="10.0.5" />
+		<PackageReference Include="System.Net.Http.Json" Version="10.0.8" />
+		<PackageReference Include="System.Text.Json" Version="10.0.8" />
 	</ItemGroup>
 	
 </Project>
