--- SampleMAUI/App.xaml.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace SampleMAUI
-{
-    public partial class App : Application
-    {
-        public App()
-        {
-            InitializeComponent();
-        }
-
-        protected override Window CreateWindow(IActivationState? activationState)
-        {
-            return new Window(new AppShell());
-        }
-    }
-}
