--- SampleMAUI/AppShell.xaml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<Shell
-    x:Class="SampleMAUI.AppShell"
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:SampleMAUI"
-    Title="SampleMAUI">
-
-    <ShellContent
-        Title="Home"
-        ContentTemplate="{DataTemplate local:MainPage}"
-        Route="MainPage" />
-
-</Shell>
