--- SampleMAUI/AppShell.xaml.cs
@@ -1,10 +1,0 @@
-﻿namespace SampleMAUI
-{
-    public partial class AppShell : Shell
-    {
-        public AppShell()
-        {
-            InitializeComponent();
-        }
-    }
-}
