--- SampleMAUI/MainPage.xaml
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:SampleMAUI"
-             x:DataType="local:MainViewModel"
-             x:Class="SampleMAUI.MainPage">
-
-    <ScrollView>
-        <VerticalStackLayout Padding="30,0" >
-            <Label>Organisation/Collection Name</Label>
-            <Entry Text="{Binding Name}"/>
-            <Label>Server Type</Label>
-            <Picker ItemsSource="{Binding ServerTypes}" SelectedItem="{Binding ServerType}" />
-            <Label>PAT</Label>
-            <Entry  Text="{Binding PAT}" />
-            <Label>Project Name</Label>
-            <Entry  Text="{Binding ProjectName}" />
-            <Label>Repository Name</Label>
-            <Entry  Text="{Binding RepoName}" />
-            <Label>Pipeline Name</Label>
-            <Entry  Text="{Binding PipelineName}" />
-            <Button Margin="10" Style="{DynamicResource AccentButtonStyle}" Command="{Binding ConnectCommand}" Text="Connect" />
-            
-            <!-- 
-                <ProgressBar Visibility="{Binding IsBusy, Converter={StaticResource booleanToVisibilityConverter}}" IsIndeterminate="True"/>
-            -->
-        </VerticalStackLayout>
-    </ScrollView>
-
-</ContentPage>
