--- SampleMAUI/Platforms/Android/MainActivity.cs
@@ -1,11 +1,0 @@
-﻿using Android.App;
-using Android.Content.PM;
-using Android.OS;
-
-namespace SampleMAUI
-{
-    [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, LaunchMode = LaunchMode.SingleTop, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
-    public class MainActivity : MauiAppCompatActivity
-    {
-    }
-}
