--- SampleMAUI/Platforms/Android/MainApplication.cs
@@ -1,16 +1,0 @@
-﻿using Android.App;
-using Android.Runtime;
-
-namespace SampleMAUI
-{
-    [Application]
-    public class MainApplication : MauiApplication
-    {
-        public MainApplication(IntPtr handle, JniHandleOwnership ownership)
-            : base(handle, ownership)
-        {
-        }
-
-        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
-    }
-}
