--- SampleMAUI/Platforms/iOS/AppDelegate.cs
@@ -1,10 +1,0 @@
-﻿using Foundation;
-
-namespace SampleMAUI
-{
-    [Register("AppDelegate")]
-    public class AppDelegate : MauiUIApplicationDelegate
-    {
-        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
-    }
-}
