--- SampleMAUI/Resources/Raw/AboutAssets.txt
@@ -1,15 +1,0 @@
-﻿Any raw assets you want to be deployed with your application can be placed in
-this directory (and child directories). Deployment of the asset to your application
-is automatically handled by the following `MauiAsset` Build Action within your `.csproj`.
-
-	<MauiAsset Include="Resources\Raw\**" LogicalName="%(RecursiveDir)%(Filename)%(Extension)" />
-
-These files will be deployed with your package and will be accessible using Essentials:
-
-	async Task LoadMauiAsset()
-	{
-		using var stream = await FileSystem.OpenAppPackageFileAsync("AboutAssets.txt");
-		using var reader = new StreamReader(stream);
-
-		var contents = reader.ReadToEnd();
-	}
