+++ SampleWPF/MainWindow.xaml
@@ -1,0 +1,39 @@
+﻿<Window x:Class="SampleWPF.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:SampleWPF"
+        d:DataContext="{d:DesignInstance Type=local:MainViewModel}"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+    <Window.Resources>
+        <BooleanToVisibilityConverter x:Key="booleanToVisibilityConverter" />
+    </Window.Resources>
+    <Grid Margin="10">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <GroupBox Header="Connection">
+            <StackPanel>
+                <TextBlock>Organisation/Collection Name</TextBlock>
+                <TextBox Text="{Binding Name, UpdateSourceTrigger=PropertyChanged}"/>
+                <TextBlock>Server Type</TextBlock>
+                <ComboBox ItemsSource="{Binding ServerTypes}" SelectedItem="{Binding ServerType}" />
+                <TextBlock>PAT</TextBlock>
+                <TextBox  Text="{Binding PAT, UpdateSourceTrigger=PropertyChanged}" />
+                <TextBlock>Project Name</TextBlock>
+                <TextBox  Text="{Binding ProjectName, UpdateSourceTrigger=PropertyChanged}" />
+                <TextBlock>Repository Name</TextBlock>
+                <TextBox  Text="{Binding RepoName, UpdateSourceTrigger=PropertyChanged}" />
+                <TextBlock>Pipeline Name</TextBlock>
+                <TextBox  Text="{Binding PipelineName, UpdateSourceTrigger=PropertyChanged}" />
+                <Button Margin="10" Style="{DynamicResource AccentButtonStyle}" Command="{Binding ConnectCommand}">Connect</Button>
+                <ProgressBar Visibility="{Binding IsBusy, Converter={StaticResource booleanToVisibilityConverter}}" IsIndeterminate="True"/>
+            </StackPanel>
+        </GroupBox>
+
+        
+    </Grid>
+</Window>
