+++ azure-pipelines-mergetest.yml
@@ -16,7 +16,7 @@ variables:
 - name: buildConfiguration
   value: 'Release'
 - name: netVersion
-  value: '10.0.105'
+  value: '10.0.107'
 
 steps:
