+++ global.json
@@ -1,7 +1,7 @@
 {
     "sdk": {
-        "version": "10.0.105",
-        "workloadVersion": "10.0.105",
+        "version": "10.0.107",
+        "workloadVersion": "10.0.107",
         "allowPrerelease": false,
         "rollForward": "disable"
     }
